refactor: move react client to root
