--- examples/svelte/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@rocketrpc/svelte-example",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
