+++ package.json
@@ -35,7 +35,7 @@
   "homepage": "https://github.com/akash-joshi/functions-over-websockets#readme",
   "dependencies": {
     "socket.io": "^4.0.0",
-    "socket.io-client": "^4.0.0"
+    "socket.io-client": "^4.3.0"
   },
   "devDependencies": {
     "ts-jest": "^29.0.5",
