+++ src/types.d.ts
@@ -1,0 +1,5 @@
+export type FunctionCallParams = {
+  id: string;
+  procedurePath: string;
+  params: any[];
+};
